Calculate averages unless insufficient data
Also simplifies the flow; necessary for correct % rolling averages
